+++ lib/src/track/options.dart
@@ -88,9 +88,13 @@ class CameraCaptureOptions extends VideoCaptureOptions {
         'facingMode':
             cameraPosition == CameraPosition.front ? 'user' : 'environment'
     };
-    if (deviceId != null) {
+    if (deviceId != null && deviceId!.isNotEmpty) {
       if (kIsWeb) {
-        constraints['deviceId'] = deviceId;
+        if (isChrome129OrLater()) {
+          constraints['deviceId'] = {'exact': deviceId};
+        } else {
+          constraints['deviceId'] = {'ideal': deviceId};
+        }
       } else {
         constraints['optional'] = [
           {'sourceId': deviceId}
@@ -331,9 +335,13 @@ class AudioCaptureOptions extends LocalTrackOptions {
       }
     }
 
-    if (deviceId != null) {
+    if (deviceId != null && deviceId!.isNotEmpty) {
       if (kIsWeb) {
-        constraints['deviceId'] = deviceId;
+        if (isChrome129OrLater()) {
+          constraints['deviceId'] = {'exact': deviceId};
+        } else {
+          constraints['deviceId'] = {'ideal': deviceId};
+        }
       } else {
         constraints['optional']
             .cast<Map<String, dynamic>>()